Moved code around, conform JobSchedule to MutableCollection
